# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SkillIdentityResponse": {
        "description": "Logical skill identity response.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentSkillVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "status",
          "current_version",
          "created_at",
          "updated_at"
        ],
        "title": "SkillIdentityResponse",
        "type": "object"
      },
...
      "SkillVersionCreateRequest": {
        "additionalProperties": false,
        "description": "Normalized JSON publish contract.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillVersionContentRequest"
          },
          "governance": {
            "$ref": "#/components/schemas/SkillGovernanceRequest"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillVersionMetadataRequest"
          },
          "relationships": {
            "$ref": "#/components/schemas/SkillVersionRelationshipsRequest"
          },
          "slug": {
            "description": "Stable public slug for the skill identity.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "description": "Immutable semantic version being published.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "content",
          "metadata"
        ],
        "title": "SkillVersionCreateRequest",
        "type": "object"
      },
...
      "SkillVersionResponse": {
        "description": "Normalized exact immutable version metadata response.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillContentSummaryResponse"
          },
          "content_download_path": {
            "title": "Content Download Path",
            "type": "string"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillMetadataResponse"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "relationships": {
            "$ref": "#/components/schemas/SkillVersionRelationshipsResponse"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          }
        },
        "required": [
          "slug",
          "version",
          "version_checksum",
          "content",
          "metadata",
          "lifecycle_status",
          "trust_tier",
          "relationships",
          "published_at",
          "content_download_path"
        ],
        "title": "SkillVersionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/skill-versions": {
      "post": {
        "description": "Create a new immutable skill version from a normalized JSON payload containing markdown content, structured metadata, and authored relationships.",
        "operationId": "createSkillVersion",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "content": {
                  "raw_markdown": "# Python Lint\n\nLint Python files consistently.\n",
                  "rendered_summary": "Lint Python files consistently."
                },
                "governance": {
                  "provenance": {
                    "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                    "repo_url": "https://github.com/example/skills",
                    "tree_path": "skills/python.lint"
                  },
                  "trust_tier": "internal"
                },
                "metadata": {
                  "description": "Linting skill",
                  "headers": {
                    "runtime": "python"
                  },
                  "inputs_schema": {
                    "type": "object"
                  },
                  "maturity_score": 0.9,
                  "name": "Python Lint",
                  "outputs_schema": {
                    "type": "object"
                  },
                  "security_score": 0.95,
                  "tags": [
                    "python",
                    "lint"
                  ],
                  "token_estimate": 128
                },
                "relationships": {
                  "conflicts_with": [],
                  "depends_on": [
                    {
                      "markers": [
                        "linux",
                        "gpu"
                      ],
                      "optional": true,
                      "slug": "python.base",
                      "version_constraint": ">=1.0.0,<2.0.0"
                    }
                  ],
                  "extends": [
                    {
                      "slug": "python.base",
                      "version": "1.0.0"
                    }
                  ],
                  "overlaps_with": [
                    {
                      "slug": "python.format",
                      "version": "1.0.0"
                    }
                  ]
                },
                "slug": "python.lint",
                "version": "1.2.3"
              },
              "schema": {
                "$ref": "#/components/schemas/SkillVersionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "content": {
                    "checksum": {
                      "algorithm": "sha256",
                      "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                    },
                    "rendered_summary": "Lint Python files consistently.",
                    "size_bytes": 123
                  },
                  "content_download_path": "/skills/python.lint/versions/1.2.3/content",
                  "lifecycle_status": "published",
                  "metadata": {
                    "description": "Linting skill",
                    "headers": {
                      "runtime": "python"
                    },
                    "inputs_schema": {
                      "type": "object"
                    },
                    "maturity_score": 0.9,
                    "name": "Python Lint",
                    "outputs_schema": {
                      "type": "object"
                    },
                    "security_score": 0.95,
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "token_estimate": 128
                  },
                  "provenance": {
                    "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                    "repo_url": "https://github.com/example/skills",
                    "tree_path": "skills/python.lint"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "relationships": {
                    "conflicts_with": [],
                    "depends_on": [
                      {
                        "selector": {
                          "markers": [
                            "linux",
                            "gpu"
                          ],
                          "optional": true,
                          "slug": "python.base",
                          "version_constraint": ">=1.0.0,<2.0.0"
                        }
                      }
                    ],
                    "extends": [
                      {
                        "selector": {
                          "markers": [],
                          "slug": "python.base",
                          "version": "1.0.0"
                        },
                        "target_version": {
                          "description": "Base runtime skill",
                          "lifecycle_status": "published",
                          "name": "Python Base",
                          "published_at": "2026-03-01T08:30:00Z",
                          "slug": "python.base",
                          "tags": [
                            "python",
                            "runtime"
                          ],
                          "trust_tier": "verified",
                          "version": "1.0.0"
                        }
                      }
                    ],
                    "overlaps_with": [
                      {
                        "selector": {
                          "markers": [],
                          "slug": "python.format",
                          "version": "1.0.0"
                        }
                      }
                    ]
                  },
                  "slug": "python.lint",
                  "trust_tier": "internal",
                  "version": "1.2.3",
                  "version_checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionResponse"
                }
              }
            },
            "description": "Immutable skill version published successfully."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DUPLICATE_SKILL_VERSION",
                    "details": {
                      "slug": "python.lint",
                      "version": "1.2.3"
                    },
                    "message": "Skill version already exists: python.lint@1.2.3"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CONTENT_STORAGE_FAILURE",
                    "message": "Failed to persist immutable skill version."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized content or metadata persistence failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish an immutable skill version",
        "tags": [
          "skills"
        ]
      }
    },
    "/skills/{slug}": {
      "get": {
        "description": "Return the normalized identity row for a skill slug.",
        "operationId": "getSkillIdentity",
        "parameters": [
          {
            "description": "Stable public slug of the skill.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-02-10T08:30:00Z",
                  "current_version": {
                    "lifecycle_status": "published",
                    "published_at": "2026-03-10T08:30:00Z",
                    "trust_tier": "internal",
                    "version": "1.2.3"
                  },
                  "slug": "python.lint",
                  "status": "published",
                  "updated_at": "2026-03-10T08:30:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillIdentityResponse"
                }
              }
            },
            "description": "Logical skill identity returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_NOT_FOUND",
                    "details": {
                      "slug": "python.lint"
                    },
                    "message": "Skill not found: python.lint"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested skill slug does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch one logical skill identity",
        "tags": [
          "skills"
        ]
      }
    },
    "/skills/{slug}/versions": {
      "get": {
        "description": "Return immutable versions in deterministic reverse chronological order.",
        "operationId": "listSkillVersions",
        "parameters": [
          {
            "description": "Stable public slug of the skill.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "slug": "python.lint",
                  "versions": [
                    {
                      "content": {
                        "checksum": {
                          "algorithm": "sha256",
                          "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                        },
                        "rendered_summary": "Lint Python files consistently.",
                        "size_bytes": 123
                      },
                      "lifecycle_status": "published",
                      "metadata": {
                        "description": "Linting skill",
                        "name": "Python Lint",
                        "tags": [
                          "python",
                          "lint"
                        ]
                      },
                      "published_at": "2026-03-10T08:30:00Z",
                      "slug": "python.lint",
                      "trust_tier": "internal",
                      "version": "1.2.3",
                      "version_checksum": {
                        "algorithm": "sha256",
                        "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                      }
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionListResponse"
                }
              }
            },
            "description": "Immutable versions listed successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_NOT_FOUND",
                    "details": {
                      "slug": "python.lint"
                    },
                    "message": "Skill not found: python.lint"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested skill slug does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List immutable versions for one skill",
        "tags": [
          "skills"
        ]
      }
    },
...
    "/skills/{slug}/versions/{version}/status": {
      "patch": {
        "description": "Transition the lifecycle state for one immutable skill version.",
        "operationId": "updateSkillVersionStatus",
        "parameters": [
          {
            "description": "Stable public slug of the skill.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to update.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to update.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillVersionStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "is_current_default": true,
                  "lifecycle_changed_at": "2026-03-11T09:15:00Z",
                  "slug": "python.lint",
                  "status": "deprecated",
                  "trust_tier": "internal",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionStatusResponse"
                }
              }
            },
            "description": "Lifecycle status updated successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "slug": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update immutable version lifecycle status",
        "tags": [
          "skills"
        ]
      }
    }
  }
}